requirements.txt
Flask==2.0.3
tensorflow-cpu==2.10.0
Pillow==10.0.0
numpy==1.24.3